bribeWebParent/bribeBackEnd/src/main/java/com/bribemap/admin/user/UserRepository.java
package com.bribemap.admin.user;

import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.jpa.repository.Query;

import org.springframework.data.repository.PagingAndSortingRepository;
import org.springframework.data.repository.query.Param;


import com.bribemap.common.entity.User;


public interface UserRepository extends PagingAndSortingRepository<User, Integer>{

	@Query("SELECT u FROM User u WHERE u.email = :email")
	public User getUserByEmail(@Param("email") String email);
	
	public Long countById(Integer id);
	
	@Query("SELECT u FROM User u WHERE u.firstName LIKE %?1% OR u.lastName LIKE %?1%")
	public  Page<User> findAll(String keyword, Pageable pageable);
}
